pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "viz2psy"
version = "0.1.0"
description = "Visual image psychological feature extraction pipeline"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "Pillow>=9.0",
    "pandas>=1.5",
    "tqdm>=4.60",
    "resmem>=1.1",
    "open-clip-torch>=2.24",
    "scikit-image>=0.21",
    "deepgaze-pytorch>=1.0",
]

[project.optional-dependencies]
dev = ["pytest", "ipython"]

[tool.setuptools.packages.find]
where = ["src"]
